+++ src/vaultknox/__init__.py
@@ -1,10 +1,11 @@
-from vaultknox.autonomous_secrets import AutonomousSecretsError, AutonomousSecretsStore
+from vaultknox.agent_guide import TRIGGERS, check_triggers, get_system_prompt_snippet
 from vaultknox.health import VaultHealthChecker
 from vaultknox.hermes_tool import vault_tool
+from vaultknox.hooks import handle
 from vaultknox.rotation import rotate_master_key
 from vaultknox.scanner import SecretScanner
-from vaultknox.verifier import CredentialVerifier
 from vaultknox.vault import VaultError, VaultKnox
+from vaultknox.verifier import CredentialVerifier
 
 __all__ = [
     "VaultKnox",
@@ -16,4 +17,8 @@ __all__ = [
     "SecretScanner",
     "CredentialVerifier",
     "VaultHealthChecker",
+    "TRIGGERS",
+    "check_triggers",
+    "get_system_prompt_snippet",
+    "handle",
 ]